branched out the filter tests into a new file
